# packages/contracts/openapi/atlas-registry-v1.yaml
openapi: 3.1.0
info:
  title: Atlas Registry API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
  /skills:
    get:
      operationId: searchSkills
      parameters:
        - { name: q, in: query, schema: { type: string, maxLength: 200 } }
        - { name: category, in: query, schema: { type: string, maxLength: 80 } }
        - { name: page, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: size, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
      responses:
        '200': { description: Access-filtered catalog result }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /skills/{slug}:
    get:
      operationId: getSkill
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string, maxLength: 100 } }
      responses:
        '200': { description: Skill detail }
        '404': { $ref: '#/components/responses/NotFound' }
  /installations/operations:
    post:
      operationId: createLifecycleOperation
      parameters:
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 8, maxLength: 160 } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOperation' }
      responses:
        '200': { description: Created or replayed deterministic operation }
        '409': { description: Idempotency conflict }
  /operations/{operationId}:
    get:
      operationId: getOperation
      parameters:
        - { name: operationId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Durable operation state }
        '404': { $ref: '#/components/responses/NotFound' }
  /registrations:
    post:
      operationId: createRegistration
      responses:
        '200': { description: Registration pending external validation }
  /skills/{skillId}/feedback:
    post:
      operationId: submitFeedback
      parameters:
        - { name: skillId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Feedback accepted }
        '400': { description: Invalid or unknown fields }
components:
  schemas:
    CreateOperation:
      type: object
      additionalProperties: false
      required: [operationType, skillId, hostCode, scope]
      properties:
        operationType: { type: string, enum: [INSTALL, UPDATE, MANUAL_ROLLBACK, UNINSTALL] }
        skillId: { type: string, format: uuid }
        installationId: { type: [string, 'null'], format: uuid }
        hostCode: { type: string }
        targetVersion: { type: [string, 'null'] }
        riskAcknowledgement: { type: [string, 'null'] }
        scope:
          type: object
          additionalProperties: false
          required: [type]
          properties:
            type: { type: string, enum: [PERSONAL, PROJECT] }
            projectRef: { type: [string, 'null'] }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, correlationId, fieldErrors]
          properties:
            code: { type: string }
            message: { type: string }
            correlationId: { type: string }
            fieldErrors: { type: array, items: { type: object } }
  responses:
    Unauthorized: { description: Authentication required }
    NotFound: { description: Resource unavailable or hidden }